changed file structure within the server file, db file was getting confused. Changed to middleware and controllers within the server folder. Added new end point, GET companyList which delivers the full list of companies from the db to the front end. This route will be used when the dashboard is initially loaded. Saving the company list in the store will give the frontend the ability to very quickly rerender the pie chart component as a user selects different companies from the list. This method, delivering the whole list at once when teh dashboard loads, is currently percieved as preferable to hitting the database every time the user wants to view the details for another company and will hopefully preform to our expecations. However, if for some reason having all that data in the store is not ideal, these routes can be refactored to deliver specific pieces of data.
