+++ src/server/controllers/company.ts
@@ -30,7 +30,7 @@ export class CompanyRepository {
    // TODO SOLVE ANY/OBJECT TYPING***********************************************
    // add is asyncronous because I cannot return the db.none method the way that I did with the user controllers
    // this async wrapper here makes sure all the data gets put into the database before we move on with the middleware
-   async add (companyData: any[]) {
+  async add (companyData: any[]) {
     for (let i = 0; i < companyData.length; i += 1) {
 
       // define the issue object which get added to the company object
@@ -47,12 +47,14 @@ export class CompanyRepository {
       // after each issues object is created, submit the issues object and the rest of the company data 
       this.db.none('INSERT INTO company (id, ticker, name, blurb, logo, issues) VALUES ($1, $2, $3, $4, $5, $6);', 
       [v4(), companyData[i].ticker, companyData[i].companyName, companyData[i].blurb, companyData[i].link, JSON.stringify(issues)])
-      .then(() => {
-        console.log('DATA ENTERED');
-      })
       .catch((error: any) => {
         console.log('ERROR AT ADD FUNCTION IN COMPANY.TS', error);
       })
     }
   };
+
+  // query to get all companies out of the db
+  getList() {
+    return this.db.any('SELECT * FROM company;')
+  }
 }
