+++ src/server/server.ts
@@ -20,9 +20,9 @@ import * as bodyParser from 'body-parser';
 // import cors to enable cross-origin sharing between dev server and server
 import * as cors from 'cors';
 // import authentication middleware
-import Authenticate from './db/controllers/authenticate'
+import Authenticate from './middleware/authenticate'
 // import companyDb middleware
-import CompanyDatabase from './db/controllers/addCompanyDataToDb';
+import CompanyDatabase from './middleware/companyDataMethods';
 
 // activate the express server
 const app: Application = express();
@@ -49,16 +49,25 @@ app.use(bodyParser.urlencoded({ extended: true }));
 app.post('/login', 
   Authenticate.compareHash,
   (_: Request, res: Response) => {
-  console.log('LOGIN SUCCESS');
-  res.end();
-});
+    res.sendStatus(200);
+  }
+);
 
 // registration end point
 app.post('/register', 
   Authenticate.hashPassword,
   (_: Request, res: Response) => {
-    console.log('USER REGRISTRATION SUCCESS');
-    res.send('USER REGISTRATION SUCCESS');
+    res.sendStatus(200);
+  }
+);
+
+// TODO UPDATE END POINT TO GET USER ISSUES
+
+// end point for deliverying a list of companies on dashboard render
+app.get('/companyList', 
+  CompanyDatabase.getCompanyList,
+  (_: Request, res: Response) => {
+    res.send(res.locals.companyDataArray);
   }
 );
 
@@ -68,7 +77,8 @@ app.post('/companyData',
   CompanyDatabase.insertData,
   (__dirname: Request, res: Response) => {
     res.sendStatus(200);
-})
+  }
+);
 
 // wake up the server
 app.listen(PORT, () => console.log(`Server listening on PORT: ${PORT}`));
